requirements.txt
beautifulsoup4==4.8.0
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.6
django-active-link==0.1.6
django-allauth==0.39.1
django-ckeditor==5.7.1
django-cleanhtmlfield==1.0.2
django-crispy-forms==1.7.2
django-filter==2.1.0
django-html-field==0.1.3
django-jet==1.0.8
django-js-asset==1.2.2
django-pagination==1.0.7
django-tinymce==2.8.0
django-tinymce4-lite==1.7.5
django-tinymce4-widget==4.0.0
django-widget-tweaks==1.4.5
google-api-python-client==1.4.1
gunicorn==19.9.0
httplib2==0.13.0
idna==2.8
jsmin==2.2.2
mysqlclient==1.4.2
oauth2client==4.1.3
oauthlib==3.0.2
Pillow==6.1.0
psycopg2==2.8.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
simplejson==3.16.0
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
